pipeline/requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
PyYAML==6.0
requests==2.28.1
urllib3==1.26.12
bs4
lxml
tqdm
nltk
